+++ src/constants/main.ts
@@ -5,6 +5,7 @@ export const extensionName = 'gruvbox-material-theme';
 export const defaultThemeOptions: ThemeOptions = {
   contrast: 'medium',
   cursor: 'white',
+  enableSparators: false,
   palette: 'material',
   selection: 'grey',
 };