+++ src/helpers/theme/get-theme-options.ts
@@ -3,6 +3,7 @@ import { defaultThemeOptions, extensionName } from '../../constants/main';
 import {
   type CursorColor,
   type SelectionColor,
+  type Separators,
   type ThemeContrast,
   type ThemeOptions,
   type ThemePalette,
@@ -15,10 +16,12 @@ export function getThemeOptions(): ThemeOptions {
   const cursor = config.get<CursorColor>('cursor');
   const palette = config.get<ThemePalette>('palette');
   const selection = config.get<SelectionColor>('selection');
+  const enableSparators = config.get<Separators>('separators');
 
   return {
     contrast: contrast ?? defaultThemeOptions.contrast,
     cursor: cursor ?? defaultThemeOptions.cursor,
+    enableSparators: enableSparators ?? defaultThemeOptions.enableSparators,
     palette: palette ?? defaultThemeOptions.palette,
     selection: selection ?? defaultThemeOptions.selection,
   };
@@ -28,6 +31,7 @@ export function checkIsDefaultThemeOptions(options: ThemeOptions) {
   return (
     options.contrast === defaultThemeOptions.contrast &&
     options.cursor === defaultThemeOptions.cursor &&
+    options.enableSparators === defaultThemeOptions.enableSparators &&
     options.palette === defaultThemeOptions.palette &&
     options.selection === defaultThemeOptions.selection
   );